% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calibration_coefs}
\alias{calibration_coefs}
\title{Calibration Coefficients}
\usage{
calibration_coefs(x, y, log = FALSE)
}
\arguments{
\item{x}{axis input values from plot}

\item{y}{axis output values from image}

\item{log}{logical indicating if axis is logged}
}
\value{
numeric vector of length 2
}
\description{
Calibration Coefficients
}
