% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolp.R
\name{tolp}
\alias{tolp}
\title{tolp}
\usage{
tolp()
}
\value{
data.frame with columns g, x, and y containing tolped data points
}
\description{
Run a shiny app to tolp data points from a plot image.
}
\examples{
\dontrun{
my_tolped_data <- tolp()
}
}
