% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tolp_resource}
\alias{tolp_resource}
\title{TOLP Resource}
\usage{
tolp_resource(...)
}
\arguments{
\item{...}{character vectors, specifying subdirectory and file(s) within the
tolp package. The default, none, returns the root of the package. Wildcards
are not supported.}
}
\value{
the contents of the resource as a string
}
\description{
TOLP Resource
}
