% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tolp_scale}
\alias{tolp_scale}
\title{Scale}
\usage{
tolp_scale(x, b, log = FALSE, reverse = FALSE)
}
\arguments{
\item{x}{input values}

\item{b}{calibration coefficients}

\item{log}{logical indicating if axis is logged}

\item{reverse}{logical indicating to go from plot to image values (FALSE)
or image to plot values (TRUE)}
}
\value{
numeric vector with same length as x
}
\description{
Scale
}
